Configure ceph mon info on 2nd region compute host
compute hosts go to reboot loop after configuring nova_local
with remote backing. The reboot was due to nova go enabled test
failed. Because the ceph monitor info was not configured in
ceph.conf on the compute node, nova-compute process failed to
start due to error connecting to the cluster.

This update configured the ceph mon info on the compute hosts in a
secondary region when the cinder is configured as a shared service
and the region has ceph backend configured. The ceph mon info is
retrieved from the primary.

Change-Id: Ie37fa1126f30bae5318377c8e1d75acc8fa04be1