Fix upgrade with Host TPM
This commit allows TPM certificates, per controller hosts, to be
persisted across:
- Host Reinstalls
- Controller Restores
- Doing a Backup on controller-1 and doing a Restore on controller-0
- DX / SX Upgrades

By populating them in Sysinv's per host tpmdevice table's tpm_data
field, we manage to push these certs in as hiera data and recreate the
TPM certs

The TPM binary memory maps are base64 encoded to allow RPC conduction
and storage in DB, as by default when the Agent RPC message is parsed,
it is done so using json.dumps which expects utf-8 encoding. The Binary
maps are base64 decoded prior to be written to the hiera records so that
the right content ends up in the files when the config manifest applies.

Change-Id: Ie8b282808afacbba92949eefb85e96d6be441822
Story: 2002886
Task: 22847
Signed-off-by: Jack Ding <jack.ding@windriver.com>