AIO-DX: Controller-1 unlock failed on first attempt after upgrade
The sysinv data is not migrated to the N+1 version until after the
sysinv agent is run.
This prevents the sysinv agent from adding the UUID to the platform
conf, which prevents compute huge from setting the correct Kernel boot
arguments.

We set the UUID during the data migration so compute huge can update the
boot arguments when the node is unlocked which will prevent the extra
reboot.

Change-Id: Iecf03dd9e30b37419939f50677305066841a2fa8
Story: 2002886
Task: 22847
Signed-off-by: Jack Ding <jack.ding@windriver.com>