Delay starting compute services until controller services are up
On AIO-SX startup there are many services that need to get started up.
We have noted experimentally that there can be significant CPU
congestion if the controller services are still initializing when we
start up the compute services.

With this change we will wait for the "cloud-services" service-group to
be active before starting up the compute services.

This has minimal effect on overall startup times because the controller
services can continue to run on all host CPUs until the cloud services
start up, at which point all host services are restricted to the
platform
CPUs.

Change-Id: Ie9f129fb078e19883cab910b3f411813237d376b