Move bindings away from controller-1 during upgrade
After the upgrade is started, controller-1 is locked, which causes
routers and DHCP from being migrated away from controller-1.
Because the database is mirrored before this point, any changes to
these bindings are lost after swacting to newer load.  This results
in a loss of connectivity over an AIO upgrade.

As a long-term solution, all these changes to the neutron DB should be
kept over the upgrade.

As an interim solution, this commit adds a migration script to manually
move the bindings off of controller-1.

Note, this script will have no effect unless compute services ar running
on controller-1.

Change-Id: I5a42a1f1f20b52665805ce73b56c383a4497b58e
Story: 2002886
Task: 22847
Signed-off-by: Jack Ding <jack.ding@windriver.com>