Reduce neutron workers from eng_workers to eng_workers_by_2
In the case of a LO DOR with neutron workers equal to the number of
cores, neutron-server was observed to consume nearly all of the CPU,
causing critical CPU alarms, and starving out the system.

In order to prevent this, this commit changes both the number of API
workers, and the number of RPC workers, to be eng_workers_by_2 rather
than eng_workers.
By cutting the number of workers in half, this change prevents too much
throughput in neutron from grinding the system to a halt, and provides a
reduction in the overall time for a LO DOR. as well as giving somewhat
more consistent DOR results, consistenly <~45 mins with this change,
rather than frequently over an hour, or not completing without this
change.

Change-Id: Ib40257ef34766a250b70edcb704220675ed0cc85