Fix Distributed Cloud Subcloud Compute Node Unlock
Service tokens were introduced in Nova Compute to resolve another
issue. The change incorrectly adds the call to the nova::keystone::auth
class, which will attempt to test & set the Nova User and the Nova
endpoints. This is a redundant operation as the Nova Controller
manifest would already have done this.

This increases the apply time of the Nova Compute manifest, since the
keystone::auth is a relatively heavy class - doing a number of
List/GET calls to enforce the idempotentency of the underlying Exec
class.

Functionally, this breaks the Compute node in a subcloud as it uses an
incorrect keystone auth url to try to create the Nova user.

Change-Id: I1809b62bbb60a04d08c0afee87945fff4c0a99a8
Signed-off-by: Tyler Smith <tyler.smith@windriver.com>