Fixes for H/W replacement during Upgrades
Disallow host-delete during upgrade:
Prior to the controller upgrade, disallow host-delete during an
upgrade.
Issue was triggered when a host-delete was allowed when the active
controller has not yet been upgraded and a host-delete was performed.
This updates sysinv to disallow the host-delete prior to controller
upgrade. The user would need to abort and delete prior to continuing.

Disallow host-lock compute during upgrade:
host-lock of computes is prevented during an upgrade as instance
migrations are not allowed prior to upgrading both controllers.
In AIO-DX, allow host-lock controller-0 after upgrading controller-1.

Story: 2002886
Task: 22847

Change-Id: If74480e9e253528ebaf0e3bc3132b93fb7aa5e8e
Signed-off-by: Jack Ding <jack.ding@windriver.com>