Restore proper iscsi target config file
Using the upstream implementation of cinder backup and restore
breaks bringing up any instance that was booted from volume
and still running at the time of the backup.

This is due to cinder restore process, writing data through iscsi
back to the volumes, thus adding new iscsi options for the
connection and breaking the existing connections to VMs.

Changes done to fix this:
- at restore, restore the iscsi config file, but also make a copy
  of it
- after volumes are restores, use 'config_controller --restore-complete'
  to overwrite the broken config file, with the original one from
  the above copy
- 'config_controller --restore-complete' replaces 'config_controller
  --restore-compute' on AIO setups
- 'config_controller --restore-complete' must now be also run on
  non-AIO setups.
- as part of the restore procedure, compute hosts cannot be unlocked
  until 'config_controller --restore-complte' is run

Story: 2002824
Task: 22741
Depends-On: https://review.openstack.org/584492

Change-Id: I441997019987d593872c82dd62baafbb397a97b1
Signed-off-by: Jack Ding <jack.ding@windriver.com>