Add checks to system service-parameter-add to look for existing entries
The service parameter table was updated under to add support for
custom service parameters with --personality and --resource options.
With that update, the uniqueness constraint was extended to include
the new resource and personality fields. Unfortunately, it seems
postgres will ignore the constraint if any of these fields are NULL.
The sysinv code relies on this uniqueness constraint for the "already
exists" error, and so any parameters that don't set all fields will slip
past the check. If all fields are set, the uniqueness constraint blocks
the add, as expected.

This update adds a query to the service-parameter-add command to check
the database to see if the entry already exists before adding it.

Story: 2002858
Task: 22807

Change-Id: I63c900a11a779e09601e6e7762fc5af3cba3c160
Signed-off-by: Jack Ding <jack.ding@windriver.com>