Add sysinv commands related to helm chart overrides
When dealing with "system" helm charts (like those used for OpenStack)
we want to support both system-generated overrides as well as
user-specified overrides.

There are various ways in helm to specify overrides, so we try to support
as many of them as possible here.

Because we're not actually upgrading a chart in helm it becomes very
difficult to reliably handle the equivalent of helm's "--set-string"
operation so for now we won't be able to provide it.

Story: 2002876
Task: 22831

Change-Id: I27d7e61b1ae0a849ecf99afbdbbaf7ba3bfcaee2
Signed-off-by: Jack Ding <jack.ding@windriver.com>