Add support for helm chart override namespaces
There are some charts which we know will be installed multiple times in
different namespaces, and we may require separate system and user
overrides depending on the namespace.

Accordingly, we need to change over to using a "chart_name/namespace"
tuple instead of just "chart_name" to locate a given set of overrides.

This updates the DB and all APIs to use the name/namespace tuple,
replaces do_helm_chart_list() with do_helm_override_list() for
consistency, and adds a list of known user override namespaces to the
list of known charts.

Story: 2002876
Task: 22831

Change-Id: I538893bae6644e158404cfc5e94c470991df024d
Signed-off-by: Jack Ding <jack.ding@windriver.com>