Add support for helm chart system overrides
Add support for querying the system overrides from the sysinv backend.
This includes querying the list of supported charts/namespaces as well
as the system overrides for a specific chart/namespace.

The guts of the override merge code is pulled into a helper so it can
be used for merging system and user overrides as well as updating
user overrides.

Change-Id: I27b25fe3800e9350c9db80a665ffba43228f315c
Signed-off-by: David Sullivan <david.sullivan@windriver.com>
Story: 2002876
Task: 22831