Adjust who gets admin privs in kubernetes
We want to give sysinv and wrsroot "admin" access in kubernetes.
Accordingly, the permissions on the auto-generated
/etc/kubernetes/admin.conf file need to be adjusted, and the KUBECONFIG
environment variable needs to be set to point to the system-wide admin
config.  (To simplify things, we will set this env variable for anyone
who can read the admin config file.)

Change-Id: I9085fd1840d99674e515436f435d6f94617b42c7
Signed-off-by: David Sullivan <david.sullivan@windriver.com>
Story: 2002876
Task: 22831