Fix return format of vcpu pin set
This update prevents mangling of the calculated vcpu pin set string
value, when it gets processed by Puppet compute manifest, to set the
value for nova.conf vcpu_pin_set.

The mangling occurs when a single core "VMs" function is configured.
For example, if the string value resulting from database reading is
'7,15', it ends up as '715' in nova.conf. This could result in
VM fails to launch.

This update prevents mangling of the calculated string value, by
wrapping it in quotes.

Story: 2002872
Task: 22824

Change-Id: Ia6559fdf2a4bd01c1d169a40e70d1374062f8d0d
Signed-off-by: Jack Ding <jack.ding@windriver.com>