Fix multi-region deployment config_region failure
When configure secondary region with config_region, the command failed
with "expected Nova public url mis-matching existing public url". The
root cause is that the service_id used to retrieve the existing url is
not initialized with correct service_name and service_type (Nova), so
that the returned url is incorrect. The patch fixed this by initializing
service_id properly.

Story: 200291
Task: 23005

Change-Id: Id96cc01f22ab8adf32335f4a4843417a5adb807b
Signed-off-by: Andy Ning <andy.ning@windriver.com>