Allow vswitch_class to be configured
This commit adds a vswitch_class option to platform::vswitch::params,
which specifies which class is to be used for the virtual switch.
In order to enable OVS as the default virtual switch, this defaults
to ::platform::vswitch::ovs, but can be overwritten to allow using
a different virtual switch.

Story: 2003526
Task: 24802
Change-Id: I3a1d3eebd4220fdf342a5269f33e786241e218e0
Signed-off-by: Joseph Richard <Joseph.Richard@windriver.com>