Update puppet vswitch service dependency on hugepage mount
The existing dependency for ensuring the hugepage directory was mounted
did not ensure that it was executed before the openvswitch service was
started.  This causes an ordering problem under some configurations
since the OVS DPDK EAL init will fail since it cannot map the hugepages.

This update ensures the dependency is defined against the service rather
than the generic platform::vswitch puppet class to correct the
incomplete dependency ordering.

Story: 2003104
Task: 24862

Change-Id: I2dc9479024bd8363503058c0a9d3034b7c662a07
Signed-off-by: Matt Peters <matt.peters@windriver.com>