Fix tox pep8 errors of type E722 in sysinv
Remove the skip clause for E722.
Fix the offending errors. This should not affect the logic of the code.

E722: "Do not use bare except".

The aim here is to be as specific as possible with the exceptions
while still catching all the exceptions we would previously expect.

Story: 2002888
Task: 23161

Change-Id: I41bc36a74f997fc25dd2192e85b8867e26011c96
Signed-off-by: Mathieu Robinson <mathieu.robinson@windriver.com>