Update puppet ceilometer upgrade dependency on service identity
The existing dependency for ceilometer upgrade cannot ensure that it
was executed after all service identites for ceilometer and gnocchi
were created. This causes ceilometer upgrade fails during subcloud
configuration since an ordering problem that ceilometer service not
being created before running ceilometer upgrade.
This commit is to update the dependency against the service identity
for ceilometer and gnocchi to ensure the correct ordering.

Closes-Bug: 1791303

Change-Id: Ibb4fb0fe15ba5c046d3c03478e4a9c1bee908f35
Signed-off-by: Angie Wang <angie.wang@windriver.com>