synchronize file update in sysinv agent
This modification adds a synchronization decorator to the
file update function of the sysinv agent.
This will ensure that the operation does not happen
concurrently with the 'runtime manifest apply' or the
'agent audit' function.
Otherwise, a race condition may happen and cause
manifest corruption or failure to properly handle alarms.

Closes-Bug: #1791969

Change-Id: If9b6baf1608850bad0e3830ebe8345f7dc68ef62
Signed-off-by: Paul-Emile Element <Paul-Emile.Element@windriver.com>