Update minimal PV size to include k8s LVs
Make sure the minimally sized physical volume for the cgts-vg volume
group is sized correctly when the --kubernetes option is used with
config_controller. Include the kubernetes specific logical volumes that
will be present in a kubernetes configuration.

Fix the large disk check to make sure that this aligns with the
kickstarts/pre_disk_aio.cfg settings. Previously, the check would use
large disk filesystem sizes with a PV sized for a small disk.

Change-Id: I208893d65a2b456a9f02926bedf8cf69aa2aa5c0
Closes-Bug: #1794567
Depends-On: Ifc6d2c0a5dbb880f8e1ff73b01bc05cb8ab22855
Signed-off-by: Robert Church <robert.church@windriver.com>