Change backup partition size for small disks
Minimum backup partition size is calculated as the sum between the
sizes of the glance and database partitions, as well as a 20GB
overhead.

This fix increase the default backup size partition to 40GB to
be in line with the above calculations, considering the database
and the glance partitions are 10GB each by default.

This also increases the minimum disk requirements from 120GB
to 130GB.

Change-Id: I15396c92136562188f9cdf82afa473361cb45aeb
Closes-bug: 1793543
Signed-off-by: Stefan Dinescu <stefan.dinescu@windriver.com>