Reducing 2M hugepages to 0 did not update the display
After reducing 2M hugepages to 0 on a single numa node system,
sysinv-agent does not send memory audit report after the host is
unlocked. As a result, CLI/horizon does not display the updated
memory information.

This problem is due to the method that determines whether the hugepages
have been allocated checks if the output of /proc/sys/vm/nr_hugepages
is great than 0. The output is 0 in this case.

This update uses the compute config completed flag to ensure the
hugepages have been allocated by manifest.

Closes-Bug: #1795066

Change-Id: I5ab48fdc5864969b25f9745ac7de2f603670e095
Signed-off-by: Tao Liu <tao.liu@windriver.com>