Mtce: Add two new port definitions to mtc.ini for SM communications
In support of the HA Improvements feature maintenance is required to,
upon request, send SM a summary of maintenance's heartbeat responsiveness
during the last 20 heartbeat periods.

This update adds the required port assignments to the mtc.ini file
in support of said communications.

With this update the mtc.ini file will be updated to contain the
following entries.

  ; Communication ports between SM and maintenance
  sm_server_port = 2124 ; port sm receives mtce commands from
  sm_client_port = 2224 ; port mtce receives sm commands from

Change-Id: I05c022f7e4dcdeaea71bc0020641baa331daae57
Story: 2003576
Task: 26837
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>