2M hugepage appears to be larger than the context help
On a newly installed system, the 2M huge page value is set by
calculating 100% remaining available memory on the initial boot.
The remaining memory in subsequent boot could be less than the
initial boot, therefore the recalculated 2M Max value (shown in the
context help) can be less than the previous set 2M huge pages.

This update calculates the initial 2M huge page value using 90%
remaining memory during the initial boot. In addition, it adds a
notification to force sysinv-agent report memory when a host
is locked. This prevents the audit delay causes the memory info not
updated on time when the user attempts to modify the memory. The
subsequent audit report is ignored when the host is locked. In this
way host memory modify and host unlock semantic check use
the same memory info as each audit could provide slightly different
available memory value (depending on the activities performed on
the system).

Closes-Bug: # 1794547

Change-Id: If3884946365254fc6be16f24df03dbd090454a76
Signed-off-by: Tao Liu <tao.liu@windriver.com>