Data interface creation failing due to existing PCI interface
This commit prevents the data interface creation from failing due to
an issue involving a previously created PCI interface.

The function set_defaults() within interface.py was setting the
interface networktype to 'data' if it was originally set to None. Then,
a check was made later to see if the provider network being used in the
newly created data interface was being used by another data interface.
Even though the already created interface's ifclass was set to 
'pci-passthrough', the networktype was set to 'data' and this was
causing the data interface creation to fail since the provider network
being assigned to the data interface was already being used by the PCI 
interface.

Now, the function set_defaults() sets the networktype properly and the
semantic check for the provider network now checks the ifclass rather
than the networktype.

Closes-Bug: 1795717

Change-Id: I79cbe67fec92552e06189c3592b7a63d17f4c909
Signed-off-by: Patrick Bonnell <patrick.bonnell@windriver.com>