Modify default lldp system description
This commit modifies the default system description configuration
for the lldpd daemon to remove a product specific name and version.
Instead, the OS name and version (default used by lldpd) will be
used. For example, for a CentOS host, the system description would
look something like:

CentOS Linux 7 (Core)

For security reasons, we also pass the -k option to the
lldpd process in order to prevent the specific kernel version to
be broadcast.

Change-Id: I2a15481b95d1e37bc2f4ae053a10c3289dd2f400
Closes-Bug: #1796112
Signed-off-by: Steven Webster <steven.webster@windriver.com>