Vlan lower interface should not have class set
The interface class of a vlan lower interface was set to platform,
causing it to fail to get the address of the interface.
The class of the interface should not be set if it is used as an
underlying interface of a vlan interface.

Change-Id: I04e2c1f1606eb0582f7f8c33324ac376fb77c312
Closes-bug: 1797391
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>
(cherry picked from commit cc447124c2f7fc6b86ea3257832df71a7cb57fef)