Add Armada manifest for the stx-openstack application
As part of integrating the use of airship-armada into StarlingX we
want to add a first cut at the armada manifest for our version of
OpenStack.

For this to be useful there are a number of changes that also need
to be made in the various chart overrides, they should be coming
shortly.

We also want to lay the groundwork for other items that will end
up living in this subdirectory, hence the README and placeholder
files.

Change-Id: I73aac4121a5d46cc8647d82364a28aa6dbcd8b1c
Story: 2003908
Task: 27505
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>