sysinv: reduce minimum allowed vswitch cpu count
The vswitch CPU count semantic check is being modified to reduce the
minimum acceptable value from 1 to 0.  This is being done to accommodate
work being done to allow a vswitch to share the set of platform CPUs and
operate in a reduced CPU usage mode rather than require a dedicated set
of CPUs to run.

Story: 2004357
Task: 27957

Change-Id: Ic6d9d4b7f46a37e294c4bdcbb2948a7310956d65
Signed-off-by: Allain Legacy <allain.legacy@windriver.com>