Initial neutron helm overrides
This commit introduces support for a containerized deployment of
neutron, with ovs-dpdk running on bare metal.

This is currently subject to the following known limitations, which
have not been included as part of this submission:
 - BGP
 - SDN
 - SFC
 - SRIOV
 - PCIPT

In a regular set up, neutron interfaces are bound from sysinv, and
neutron hosts are managed from VIM. In a containerized deployment,
this (sysinv|VIM) <--> neutron interaction is nonexistent, so this
setup must be done manually, with the following commands:
 - neutron host-create
 - neutron host-bind-interface
The configured providernets, hostnames, and uuids should match the
values in sysinv.

Story: 2003909
Task: 27082
Change-Id: I3175aacfcdc96bfbc4d32fced0d63a26e7ccead0
Signed-off-by: Joseph Richard <joseph.richard@windriver.com>