sysinv pnet support without neutron extension
This commit adds support for configuring providernets in sysinv,
without communicating with neutron, in the case that the neutron
providernet extension is not supported.
This commit also adds a check in sysinv for if it is a containerized
setup, and if so, then _neutron_providernet_extension_supported
returns False.

Story: 2003909
Task: 27082
Change-Id: Id856c75b630335f5330306d165fa5205b05ce9b2
Signed-off-by: Joseph Richard <joseph.richard@windriver.com>