Support an Alternative Backup Restore Procedure
This code change is to support system backup and restore
without reinstalling storage nodes.

The config_controller command is modified as follows:

  A new option is added to system restore command
  "config_controller --restore-system <name>" with default set
  to exclude-storage-reinstall.

  config_controller --restore-system <include-storage-reinstall |
      exclude-storage-reinstall> <name>

Add a new command "tidy_storage_post_restore" to be run after
the restore. It scans Cinder/Glance and rbd backend for any
discrepancy between the two and generates a user action log file.

This code change has been tested on a virtual box as well as
hardware lab.

Story: 2004184
Task: 27672
Change-Id: I5e64fde70b977ea4bc3a5927bcbd852a393caec2
Signed-off-by: Wei Zhou <wei.zhou@windriver.com>