Adjust the ceph mon IP for simplex k8s installs
For simplex installs, the loopback interface is assigned the host's mgmt
addresses. The ceph monitor address for a host is the host's mgmt
address. A k8s pod that needs to access the external ceph cluster
requires the ceph monitor IP to be accessible from within the cluster.

Because the pod would use the host's mgmt address (i.e. a loopback
address), the ceph/rbd client within the pod will fail to communicate to
the monitor.

Move the ceph monitor to the OAM address so that it is accessible from
pods within the k8s cluster.

Change-Id: I551dc33c51c1ed42d2e558b303d67618609c6d56
Story: 2002844
Task: 26877
Signed-off-by: Robert Church <robert.church@windriver.com>