Update helm overrides for ceph on simplex installs
Consolidate the code related to obtaining formatted ceph monitor
addresses used in the chart overrides. Since this is common
functionality for the rbd_provisioner, glance, and cinder charts, move
the core functions to the helm base class.

In the case of the k8s AIO simplex, it has a single monitor that cannot
use the defined mgmt IP address on the loopback interface. The address
will be moved to use the OAM interface in the platform puppet code so
align the behavior of the helm charts as well with this update.

Change-Id: I3dbae1438e787391bc8c9f39f555cb5860d02cf7
Story: 2003909
Task: 27084
Signed-off-by: Robert Church <robert.church@windriver.com>