Update config to set values as string, rather than unicode
A prior update config for python2/3 compatibility resulted
in config values being set as unicode rather than strings. This
ended up breaking patching on IPv6 systems, as the lib,
if_nametoindex function used by patching in configuration of its
multicast addresses does not properly handle unicode.
As a result, the patching daemons were unable to communicate with
each other.

This update to tsconfig/config casts the values as str() to
restore it to the previous behaviour.

Closes-Bug:1800195

Depends-On: Ibdae0a338a1af493ac522757b61bbd6df1da941e
Change-Id: Ie7ac29728e81d2e5fcb80cde34d723c1480566d0
Signed-off-by: Sun Austin <austin.sun@intel.com>