Make LLDP OVS agent driver more robust to json output
An issue was seen recently in which a bridge with a single port caused
an exception in the LLDP driver, as the json output did not match what
was expected.  In this case, the port was a simple uuid/value list
rather than a list of uuid/value lists.

ie.

["br-ex",["uuid","val1"]]

vs

["br-int",["set",[["uuid","val2"],["uuid","val3"]]]]

This commit adds a fix to handle  both cases, and also adds
robustness to gracefully catch similar issues of unexpected json
output.

Closes-Bug: #1802118
Change-Id: I73189f138a543f34e467f51e18544b521a00457c
Signed-off-by: Steven Webster <steven.webster@windriver.com>