Add mariadb override to work around unclean shutdown
Currently on a lock/unlock the mariadb pod does not do a clean
shutdown.  Because of this, on startup it doesn't want to do an
automatic recovery.

To work around this, add an override that tells the chart to
always recover on the one mariadb pod that we have.

Once we move to a galera cluster, we'll need to change this.

Story: 2003909
Task: 27086
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>

Change-Id: I78f5dff402ad90751901f510a2ace973b7783d7f