puppet-manifests: validate cpu list before setting latencies
Do not attempt to set wakeup latency unless the list of CPU instances is
non-empty.  If either the low or high wakeup list is empty then there is
no need to attempt to run the script.  Running the script with an empty
list results in a manifest apply errors and fails the node
initialization.

Story: 2004357
Task: 27957

Change-Id: I89f19388d131b6c377020c55f5b13960738e350d
Signed-off-by: Allain Legacy <allain.legacy@windriver.com>