downgrade calico to v3.1.4 and add IP autodetect
Calico is being downgraded to v3.1.4 since the latest versions
are causing readiness and liveness failures.  Based on K8S v12.1,
Calico should be on v3.3.x, therefore the underlying issue should
be investigated further and Calico upgraded once resolved.

This update also changes the IP autodetect method to ensure the
proper IP is selected.  This is required for hosts that have
multiple IP addresses and the wrong address is being selected
based on the default find first address method.

Change-Id: I31c3630bde69160786866d0bc1bc29816892943f
Story: 2002843
Task: 22791
Signed-off-by: Matt Peters <matt.peters@windriver.com>