Fix haproxy/novnc interaction in containers setup
nova-novncproxy pod is currently listening at platform floating mgmt
address port 6080 in containers setup.  However this implicitly requires
the pod to track the active controller which is not desired or
enforceable causing the pod to frequently fail to come up.

To fix, don't override config option vnc/novncproxy_host.  This will
revert behaviour to upstream default of listening at 0.0.0.0:6080 but
also requires that we do not configure nova forwarding in haproxy to
avoid port conflict.  This is in line with overall networking
architecture for containers.

Story: 2003909
Task: 27081
Change-Id: I8eff1ad837c6a874a2a11e2f6bc73876042bb171
Signed-off-by: Gerry Kopec <gerry.kopec@windriver.com>