Remove container dependencies on sysinv-conductor shutdown
After an application apply, a docker armada container is created and
running to handle any subsequent armada commands. This container mounts
directories in /opt/platform to get access to the armada manifests and
the helm chart overrides. This running container will prevent a swact
from occurring as the DRBD resource is held open.

Update the OCF script to stop and remove the armada container after the
sysinv-conductor has been confirmed to be stopped.

Change-Id: Ia123270616a2849bb30934df09f6947023d0caba
Story: 2003908
Task: 28124
Signed-off-by: Robert Church <robert.church@windriver.com>