Fix helm overrides merge issues
This commit addresses the issues related to merging system and user
overrides. It removes the user overrides restriction (to openstack)
as well as the need to scope the overrides in values yaml file and/or
set strings.

Tests conducted:
Stx-openstack application overrides regeneration verified with the
following user overrides:
  system helm-override-update ingress kube-system \
    --set endpoints.ingress.port.http.default=8008 \
    --set pod.replicas.error_page=2
  system helm-override-update ingress openstack \
    --set pod.replicas.ingress=2
  system helm-override-update keystone openstack \
    --set conf.keystone.DEFAULT.debug=True

Story: 2003909
Task: 28246
Task: 28248

Change-Id: Ie2ada9e84721b23c8394c6e472560696a4c72fc7
Signed-off-by: Tee Ngo <Tee.Ngo@windriver.com>