Standby controller in reboot loop after modifying the partition size
The root cause of this issue is that when the newly standby controller
is coming up, the partition manifest calls "sm-unmanage / sm-manage"
for drbd-cinder service. However at this time /var/run/sm.db is not
available yet which causes "sm-unmanage / sm-manage" to fail.

This commit adds a check in the partition puppet file that only
if the system is up "sm-unmanage" and "sm-manage" will be called.

Change-Id: I1869f024579350265684f2ee3bb5e3e74e6427cb
Closes-bug: 1809009
Signed-off-by: Wei Zhou <wei.zhou@windriver.com>