Update number of replicas to reflect number of controllers
The recent change to specify "required" for anti-affinity has
highlighted the fact that heat and neutron do not properly override
the number of pod replicas.  This causes problems on AIO-SX.

Update the system overrides appropriately to set the number of
replicas based on the number of controller nodes.

Change-Id: I76769147c593974b1868ca8452d240ad391d0770
Story: 2004520
Task:  28633
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>