Kubernetes Neutron VIM Host Management Refactoring
Rework neutron system host management to operate on agent states
rather than an extended host entity, as it was agreed with the
neutron team that a new host level entity was not desired in the
neutron core.

This commit modifies the neutron override values to accommodate
the above.

Story: 2003857
Task: 26669
Depends-On: I441fcf3c186f68d17abafe337af71d0caf9c40da

Change-Id: I85943d5b6f421f4ef379694f21ad2d758bb5f38a
Signed-off-by: Kevin Smith <kevin.smith@windriver.com>