Fix missing default route on worker nodes
On the worker and storage nodes, the mgmt interface should be
set to dhcp and the cluster host alias should be set to use static IP.
With mgmt and cluster host interface sharing the same interface,
the mgmt alias interface was incorrectly set to static instead of
dhcp which causes the default route to be removed during host unlock.
This commit is to set the address method of the alias interfaces
correctly.

Story: 2004273
Task: 27826

Change-Id: I6deee76a5ea25e7753bf0f53c499922bc5d66ec6
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>