Add helm chart for galera arbitrator
As part of the work to enable clustered database, a new helm chart is
needed to define how to run the galera arbitrator.  This chart makes
use of the newly-created stx-mariadb Docker image which has the galera
arbitrator added to the upstream openstack-helm mariadb Docker image.

Since we have a new system chart, add sysinv support to handle any
system overrides.  Initially this just means specifying the Docker
image.

When we go to merge this upstream, we should try to add the galera
arbitrator to the existing upstream Docker image and the existing
mariadb helm chart.  One upstream dev has said that they'd be willing
to accept this.

Change-Id: If0362916d3b575adabf9d6c8cc467e488b249b7b
Story: 2004712
Task: 29054
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>