Ensure puppet dns config done after networking config
In some labs, the puppet manifest application during
config_controller results in the ::platform::network class
being applied after the ::platform::dns class. This causes the
/etc/resolv.conf file to be wiped (when the
apply_network_config.sh is run by the ::platform::network
class).

The fix is to introduce a dependency to ensure that the
networking config is completed before the dns config is done.

Change-Id: Ie4a8b67e5eb3594f5c21e22f8014d2b8162a0758
Related-bug: 1812449
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>