Update criteria to determine if stx-openstack is installed
The current check to determine if the stx-openstack
application is installed just checks for the APP_APPLY_SUCCESS
status. However, the application can be re-applied (either
manually or automatically when a host is unlocked) and can
then transition to the APP_APPLY_IN_PROGRESS or
APP_APPLY_FAILURE status. Updating the check to account for
these two additional status, which also indicate that the
stx-openstack application is installed.

Change-Id: I02e92eb5bcbc70c7802e4e46dc608fc08c86b1c6
story: 2002843
task: 29361
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>