use one mariadb server pod for AIO-DX
In the case of AIO-DX there are only two controllers and only one
of them will be "active" at a time.  This makes it difficult to
reliably use multiple mariadb server pods since there is no third
node to ensure the remaining nodes have quorum if one node goes
down.

Accordingly, for now we will use a single mariadb server pod for
this case.

Change-Id: Iadd72df667b09fb3533eb8f80b7fefd9414b7f07
Signed-off-by: Chris Friesen <chris.friesen@windriver.com>
Story: 2004712
Task: 29382