Fix cluster host network when mgmt is tagged
When management interface is tagged on pxeboot, the cluster-host
network should be assigned to the mgmt interface by default, not
the pxeboot interface.
This commit is to set the vlan id of the cluster-host interface
to that of the management interface if the cluster-host interface
is not specified in the configuration file.

Closes-Bug: 1815053
Change-Id: Ia9b003af118df21eecba4d1b644c2738761e7553
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>