Re-install host failed after http port changed
After changing the http port, reinstall a host will fail,
and it requires delete/add the host to recover.

This problem is due to reinstall does not update the
pxeboot config file until a unlock action is performed;
consequently the boot parameter is not updated with
the new port number.

This update sets the configure_required flag when the
host is reinstalled.

Closes-Bug: 1815495

Change-Id: Ie4f8206f93370ce4555bce9d53879b784200ba14
Signed-off-by: Tao Liu <tao.liu@windriver.com>