Fix configutilities and controllerconfig installs in DevStack
Use the DevStack-provided functions to do the Python installations
for configutilities and controllerconfig.

Prepare the plugin setting for declaring DevStack prereqs that
is available in master's DevStack playbook.

Also do not enable all services by default. sysinv-api is disabled
in the devstack job as it does not properly start under Bionic.  We
will address this separately.

Change-Id: Ib57863526d285049b5964828e1b60bf215d25a23
Signed-off-by: Dean Troyer <dtroyer@gmail.com>