Add DNS requirement for kubernetes and helm.
`helm init` is being execute before networking and DNS is properly
configured in the controller. A dependency was added to kubernetes
to setup DNS, helm manifest was updated to depend on kubernetes.

Also, the `--skip-refresh` flag was added to helm init for second
controller to avoid timeout scenarios on proxy enviroments.

Closes-Bug: 1814968

Change-Id: I65759314b3a861e7fdb428889aa5f5c1c7037661
Suggested-by: Mingyuan Qi <mingyuan.qi@intel.com>
Signed-off-by: Erich Cordoba <erich.cordoba.malibran@intel.com>