Keystone DB sync - sysinv to re-auth when receives 401
When user or project synchronization changes user ID or project ID
on subcloud, sysinv will fail with return code 401 (authentication
failure) from keystonemiddleware when it services a request. But the
failure is actually a 404 (not found) error for sysinv itself
(keystonemiddleware converted this 404 to 401 and return it back to
sysinv). The consequence is that the caller to sysinv keeps on
re-autheticate but sysinv itself never succeeds its own authentication,
until its current token expires. This commit forces sysinv to
try re-authenticate itself before it return the 401 to its caller,
thus significantly shorten sysinv service recovery time when its ID
changes.

Story: 2002842
Task: 22787

Depends-On: https://review.openstack.org/#/c/641498

Change-Id: I77a231e3eaf7057b8f76854d2f23e3ce1339f3db
Signed-off-by: Andy Ning <andy.ning@windriver.com>