Put py27 and py35 jobs back on Xenial
The upstream OpenStack default node type changed recently
from Ubuntu Xenial to Bionic. THere is additional work to
cleanly run these test on Bionic (including bumping to py36
or py37) so just go back to Xenial for now.

Added pycrypto to trigger the jobs to run since we'll need it
for the bionic bits anyway.

FWIW https://review.openstack.org/#/c/643466/ is the attempt at
switching to py37 on bionic.

Change-Id: Ib09a07fa01388b5ff984720b6290d47a01a3e6fb
Signed-off-by: Dean Troyer <dtroyer@gmail.com>
Signed-off-by: Scott Little <scott.little@windriver.com>