Increase tolerance for declaring nova services down
The nova conductor listens for heartbeats from the nova
compute services running on worker nodes. The services send
this heartbeat every 10s and the default timeout is 60s.
However, in some cases (e.g. when a controller host is powered
down unexpectedly), it can take more than a minute for the
rabbitmq server to restore messaging functionality, so
nova compute services will be declared down incorrectly. This
causes the VIM to migrate instances away from the worker
node, which we want to avoid.

To make this more tolerant of temporary failures in the
rabbitmq server, I am increasing the timeout (service_down_time)
to 90s.

Change-Id: Iedca615edf0bef8be4a05a12960071ebd055aae2
Closes-Bug: 1816848
Signed-off-by: Bart Wensley <barton.wensley@windriver.com>