enable N/Y as valiable input for Distributed Cloud Configuration
During the initial setup using config_controller, the question about
Configuring a Distributed Cloud Controller requests a [y/N], "N" is
not a valid choice, while later similar requests accepts a "N" as a
valid response.

The input should be validated in a consistent fashion if a y/N
response is requested.

Closes-bug: 1815643

Change-Id: Ib3206266d8aabe4ae025336a20485f56e1033b78
Signed-off-by: SidneyAn <ran1.an@intel.com>