Fix the image download issue if unknown image ever specified
We have a yaml file which has a section "download_images" stores
the required images. The required images list will be regenerated
when applying application. The current logic is to extend the stored
image list with the newly generated image list which causes the
application apply to always fail if the user ever specifies an
unknown image. The extend should be removed.

Change-Id: I6ee031c7f866ab213e2ef8204c1f829567dcb376
Closes-Bug: 1821344
Signed-off-by: Angie Wang <angie.wang@windriver.com>