Remove ingress http port helm override
Previously, the http port for kube-system ingress was set to 8081 in
the helm override due to conflicts with port 80 and 8080 when the oam
floating IP was used.
Since the default local domain or a configured FQDN is now used to
connect to horizon and novncproxy, the http port for ingress needs
to be set back to the default port 80 for http.

Closes-bug: 1822212

Change-Id: I253ea54294703f74b042202bf444863f3a888cf0
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>