Cleaning up panko from bare metal
All of the service parameters are removed from sysinv since
they are now driven through helm charts.

There is no panko DB or config on bare metal, so there is
no need for upgrade, backup and restore or region config
to have panko related code.

Verified that the panko service parameters are not present.

The sysinv DB schema will be cleaned up in a later commit.

Story: 2004764
Task: 30519
Change-Id: I6e6e0ab41083c6825f73bd287d3babd678e2ef64
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>