Fix kubernetes k8s-infra cpuset.cpus mangled values
The platform kubernetes cgroup configured value for cpuset.cpus
(i.e., /sys/fs/cgroup/cpuset/k8s-infra/cpuset.cpus) is sometimes
incorrect.  The cpulist values are correct in hierdata, but are
mangled when used by puppet.  Solution is to wrap the values in
quotes, like we do for various other parameters.

This bug was introduced by this code:
 https://review.openstack.org/#/c/648511/

Change-Id: Ic49090502242cc1f59dd09afad0db46cc9e399c2
Closes-Bug: 1824563
Signed-off-by: Jim Gauld <james.gauld@windriver.com>