Remove a manifest wait when setting up ovs-dpdk pods
This commit added support for wait in the pods
https://review.openstack.org/#/c/651380

However, when ovs-dpdk vswitch type is enabled like this:
  system modify --vswitch_type ovs-dpdk
the wait causes armada to timeout.

This fix is to re-comment out the wait.

Note: this causes the armada logs to show:

WARNING armada.handlers.wait [-] [chart=openvswitch]:
 "label_selector" not specified,
 waiting with no labels may cause unintended consequences.

This submission will get sanity to pass.  A later submission
by someone with ovs expertise can update the openvswitch.py
helm code to add a meta_override to eliminate the warning logs.

Partial-Bug: 1824829
Change-Id: I1e08b2dd98d859d0b37612aba3de70d969653cda
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>